# UbyTecAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "UbyTecAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Afiliados/estado/{estado}": {
      "get": {
        "tags": [
          "Afiliados"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "estado",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": { }
            }
          }
        }
      }
    },
...
    "/api/EstadoRepartidor": {
      "get": {
        "tags": [
          "EstadoRepartidor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/EstadoRepartidor"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "EstadoRepartidor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EstadoRepartidor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/EstadoRepartidor"
            }
          }
        }
      }
    },
    "/api/EstadoRepartidor/{id}": {
      "get": {
        "tags": [
          "EstadoRepartidor"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/EstadoRepartidor"
            }
          }
        }
      },
      "put": {
        "tags": [
          "EstadoRepartidor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EstadoRepartidor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "EstadoRepartidor"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Pedido": {
      "get": {
        "tags": [
          "Pedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pedido"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Pedido"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        }
      }
    },
    "/api/Pedido/{id}": {
      "get": {
        "tags": [
          "Pedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Pedido"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Pedido"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Pedido"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Pedido/cliente/{clienteId}": {
      "get": {
        "tags": [
          "Pedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "clienteId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pedido"
              }
            }
          }
        }
      }
    },
    "/api/Pedido/afiliado/{afiliadoId}": {
      "get": {
        "tags": [
          "Pedido"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "afiliadoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Pedido"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "EstadoRepartidor": {
      "type": "object",
      "properties": {
        "estadoRepartidorID": {
          "format": "int32",
          "type": "integer"
        },
        "repartidorID": {
          "format": "int32",
          "type": "integer"
        },
        "disponible": {
          "type": "boolean"
        },
        "repartidor": {
          "$ref": "#/definitions/Repartidor"
        }
      },
      "additionalProperties": false
    },
    "Pedido": {
      "type": "object",
      "properties": {
        "pedidoID": {
          "format": "int32",
          "type": "integer"
        },
        "clienteID": {
          "format": "int32",
          "type": "integer"
        },
        "afiliadoID": {
          "format": "int32",
          "type": "integer"
        },
        "repartidorID": {
          "format": "int32",
          "type": "integer"
        },
        "cantidadProducto": {
          "format": "int32",
          "type": "integer"
        },
        "nombreProducto": {
          "type": "string"
        },
        "montoProducto": {
          "format": "double",
          "type": "number"
        },
        "total": {
          "format": "double",
          "type": "number"
        },
        "nombreCliente": {
          "type": "string"
        },
        "nombreComercio": {
          "type": "string"
        },
        "nombreRepartidor": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
...
    "Repartidor": {
      "type": "object",
      "properties": {
        "repartidorID": {
          "format": "int32",
          "type": "integer"
        },
        "nombreCompleto": {
          "type": "string"
        },
        "direccionProvincia": {
          "type": "string"
        },
        "direccionCanton": {
          "type": "string"
        },
        "direccionDistrito": {
          "type": "string"
        },
        "correoElectronico": {
          "type": "string"
        },
        "usuario": {
          "type": "string"
        },
        "passwordHash": {
          "type": "string"
        },
        "telefonos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TelefonoRepartidor"
          }
        }
      },
      "additionalProperties": false
    },
...
    "TelefonoRepartidor": {
      "type": "object",
      "properties": {
        "telefonoRepartidorID": {
          "format": "int32",
          "type": "integer"
        },
        "numero": {
          "type": "string"
        },
        "repartidorID": {
          "format": "int32",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
...
  }
}